+++ src/main/kotlin/net/onedaybeard/transducers/into.kt
@@ -4,6 +4,8 @@ import java.util.concurrent.atomic.AtomicBoolean
 
 
 @Suppress("UNCHECKED_CAST")
+@Deprecated(message = "moving toward kotlin-like naming conventions",
+            replaceWith = ReplaceWith("listOf(xf, input)"))
 fun <R : MutableList<A>, A, B> MutableList<A>.into(xf: Transducer<A, B>,
                                                    input: Iterable<B>): R {
 
@@ -11,6 +13,8 @@ fun <R : MutableList<A>, A, B> MutableList<A>.into(xf: Transducer<A, B>,
 }
 
 @Suppress("UNCHECKED_CAST")
+@Deprecated(message = "moving toward kotlin-like naming conventions",
+            replaceWith = ReplaceWith("listOf(xf, input)"))
 fun <R : MutableList<A>, A, B> intoList(xf: Transducer<A, B>,
                                         init: R = mutableListOf<A>() as R,
                                         input: Iterable<B>): R {
@@ -18,6 +22,8 @@ fun <R : MutableList<A>, A, B> intoList(xf: Transducer<A, B>,
 }
 
 @Suppress("UNCHECKED_CAST")
+@Deprecated(message = "moving toward kotlin-like naming conventions",
+            replaceWith = ReplaceWith("setOf(xf, input)"))
 fun <R : MutableSet<A>, A, B> MutableSet<A>.into(xf: Transducer<A, B>,
                                                  input: Iterable<B>): R {
 
@@ -26,21 +32,10 @@ fun <R : MutableSet<A>, A, B> MutableSet<A>.into(xf: Transducer<A, B>,
 }
 
 @Suppress("UNCHECKED_CAST")
+@Deprecated(message = "moving toward kotlin-like naming conventions",
+            replaceWith = ReplaceWith("setOf(xf, input)"))
 fun <R : MutableSet<A>, A, B> intoSet(xf: Transducer<A, B>,
                                       init: R = mutableSetOf<A>() as R,
                                       input: Iterable<B>): R {
     return transduce(xf, intoRf<A, R>(), init, input)
 }
-
-private fun <A, R : MutableCollection<A>> intoRf(): ReducingFunction<R, A> {
-    return object : ReducingFunction<R, A> {
-        override fun apply(result: R,
-                           input: A,
-                           reduced: AtomicBoolean): R {
-
-            result.add(input)
-            return result
-        }
-    }
-}
-